Abstract RESPVisualized and TFSAVisualized
