+++ src/Visualized/Accounts/AccountModel.ts
@@ -1,0 +1,6 @@
+import { AccountAction } from "./ActionModel";
+
+export interface AccountInfo {
+    actions: AccountAction[],
+    yearBorn: number,
+};
