+++ src/Visualized/Accounts/RESP/RESPConstants.ts
@@ -1,6 +1,6 @@
 export const CESG_LIMIT = 7200;
 export const CESG_ABBREV = 'Canada Education Savings Grant';
-export const RESP_LIMIT = 50000;
+export const RESP_CONTRIBUTION_LIMIT = 50000;
 export const CESG_INFO_URL= 'https://www.canada.ca/en/revenue-agency/services/tax/individuals/topics/registered-education-savings-plans-resps/canada-education-savings-programs-cesp/canada-education-savings-grant-cesg.html';
 export const EAP_ABBREV = 'Educational Assistance Payment';
 export const FIRST_RESP_YEAR = 1900;
