--- src/Visualized/Accounts/RESP/RESPModel.ts
@@ -1,6 +1,0 @@
-import { AccountAction } from "../ActionModel";
-
-export interface InfoRESP {
-    actions: AccountAction[],
-    yearCreate: number
-}
