--- src/Visualized/Accounts/TFSA/TFSAModel.ts
@@ -1,6 +1,0 @@
-import { AccountAction } from "../ActionModel";
-
-export interface TFSAInfo {
-    actions: AccountAction[],
-    yearBorn: number,
-};
